+++ README.md
@@ -8,7 +8,7 @@ Implement the ability to reorder the goods with the given buttons.
     - remove `is-light` class from the button when this order is applied;
 1. `Sort by length` button should sort the goods by length.
     - remove `is-light` class from the button when this order is applied;
-1. `Reverse` button changes the current order to the opposite one. 
+1. `Reverse` button changes the current order to the opposite one.
     - it should work together with any sorting;
     - remove `is-light` class from the button when goods are reversed;
     - the second click should show goods in the direct order;
@@ -18,6 +18,5 @@ Implement the ability to reorder the goods with the given buttons.
 ## Instructions
 
 - Implement a solution following the [React task guideline](https://github.com/mate-academy/react_task-guideline#react-tasks-guideline).
-- Use the [React TypeScript cheat sheet](https://mate-academy.github.io/fe-program/js/extra/react-typescript).
 - Open one more terminal and run tests with `npm test` to ensure your solution is correct.
 - Replace `<your_account>` with your Github username in the [DEMO LINK](https://<your_account>.github.io/react_list-of-goods/) and add it to the PR description.
