+++ checklist.md
@@ -1,32 +1,4 @@
-1. [CODE KNOWLEDGE] - Don't interact with DOM directly, use React as much as possible;
-2. [CODE KNOWLEDGE] - When you rendering a list, don't forget to add `key` prop
-
-BAD EXAMPLE:
-```jsx
-<div>
-  {cats.map(cat => <Cat cat={cat} />)}
-</div>
-```
-
-GOOD EXAMPLE:
-```jsx
-<div>
-  {cats.map(cat => <Cat key={cat.id} cat={cat} />)}
-</div>
-```
-3. [UX] - don't forget to mark the active button to explain to the user what action is active
-4. [CODE STYLE] - Method's name should start with a verb (so you could easily tell what your method actually do)
-
-BAD EXAMPLE:
-```jsx
-clickHandler = () => {
- console.log('Hello, world');
-}
-```
-
-GOOD EXAMPLE:
-```jsx
-handleClick = () => {
- console.log('Hello, world');
-}
-```
+1. Don't interact with DOM directly, use React as much as possible;
+2. Add `key` to all iterations
+3. Highlight the active button
+4. Method's name should start with a verb
